backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "graphql with fastapi"
authors = ["phunguyenTPP <phu.nguyen@tppsoft.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.108.0"
uvicorn = {extras = ["standard"], version = "^0.25.0"}
sqlmodel = "^0.0.14"
strawberry-graphql = {extras = ["debug-server"], version = "^0.217.0"}
alembic = "^1.13.1"
dependency-injector = "^4.41.0"
pydantic-settings = "^2.1.0"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
